requirements.txt
tensorflow==2.16.1
numpy
pillow
matplotlib
fastapi
uvicorn
python-multipart